python/github-repo-checkout/requirements.txt
appdirs==1.4.3
astroid==2.3.2
attrs==19.3.0
black==19.3b0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
entrypoints==0.3
flake8==3.7.8
gitdb2==2.0.6
GitPython==3.0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.3
requests==2.22.0
six==1.12.0
smmap2==2.0.5
toml==0.10.0
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
